Move surface path code
